Upgraded ThyraCrsMatrixFactory to be able to create FECrsMatrix
The factory can create linear op either as FECrs or Crs matrices.
In the former case, nonlocal (w.r.t. the range vs) rows will be considered
(if any), while in the latter they will be discarded.
Since a FECrsMatrix IS a CrsMatrix, one may be tempted to not do this
distinction. However, creating a CrsMatrix ensures that a cast to
FECrsMatrix would fail, and the matrix will be treated as a CrsMatrix.
This is important in the fillComplete thyra utility, which performs
different things depending on whether the matrix is Crs or FECrs.
The alternative would have been to create an additional utility,
globalAssemble. However, it seemed easier to leave the utility as is.

Note: by default, nonlocal rows ARE considered, so that a FECrs matrix will
be created.