Using a separate function for FE begin/end fill.
FE assembly should be done only when completing a FE assembly. However, the
assign utility needs to open/close the matrix, but shoudl *NOT* perform a
global assembly.

Additionally, we need to re-init the tpetra thyra LinearOp every time we
begin/end a FE assembly, because the domain/range vector spaces need to be
updated.